requirements_general.txt
# python 3.10
customtkinter
tkPDFViewer
markdown
tkinterhtml
pillow
PyPDF2
CTkMessagebox
langchain
langchain_community
langchain_google_genai
langchain-anthropic
langchain-openai
faiss-cpu
pywebview
pythonnet
cefpython3